Module fusionner::git

Convenience wrapper around the git2-rs library

In particular, you would want to start with the git::Repository struct.

Structs

RefspecStr
Wraps around a Refspec string to provide convenience method

Remote
Wraps around a git2::Remote struct and offers convenience methods

RemoteHead
Cloned from a git2::RemoteHead without the associated lifetime. The fields correspond one to one with git2::RemoteHead.

Repository
Repository struct to wrap around git2::Repository