Struct fusionner::git::Remote

pub struct Remote<'repo> {
    pub remote: Remote<'repo>,
    // some fields omitted
}

Fields
remote: Remote<'repo>
Methods

impl<'repo> Remote<'repo>

fn disconnect(&mut self)

fn name(&self) -> Option<&str>

fn refspecs(&self) -> Refspecs

fn remote_ls(&mut self) -> Result<Vec<RemoteHead>, Error>

fn head(&mut self) -> Result<Option<String>, Error>

fn resolve_head(heads: &[RemoteHead]) -> Option<String>

fn fetch(&mut self, refspecs: &[&str]) -> Result<(), Error>

fn push(&mut self, refspecs: &[&str]) -> Result<(), Error>

fn generate_refspec(&self, src: &str, force: bool) -> Result<String, String>

For a given local reference, generate a refspec for the remote with the same path on remote i.e. refs/pulls/* --> refs/pulls/*:refs/remotes/origin/pulls/*

fn add_refspec(&self, refspec: &str, direction: Direction) -> Result<(), Error>

fn add_refspecs(&self,
refspecs: &[&str],
direction: Direction)
-> Result<(), Error>

fn resolve_target_ref(&mut self,
target_ref: Option<&str>)
-> Result<String, Error>

fn find_matching_refspec<'a>(refspecs: Refspecs<'a>,
direction: Direction,
refspec: &str)
-> Option<Refspec<'a>>

fn direction_eq(left: &Direction, right: &Direction) -> bool