Struct fusionner::git::Repository

pub struct Repository<'repo> {
    pub repository: Repository,
    // some fields omitted
}

Repository struct to wrap around git2::Repository

Examples

extern crate tempdir;
extern crate fusionner;
use fusionner::RepositoryConfiguration;
use fusionner::git::Repository;
use tempdir::TempDir;

let td = TempDir::new("checkout").unwrap();
let configuration = RepositoryConfiguration {
    uri: "https://github.com/lawliet89/fusionner.git".to_string(),
    checkout_path: td.path().to_str().unwrap().to_string(),
    fetch_refspecs: vec![],
    push_refspecs: vec![],
    username: None,
    password: None,
    key: None,
    key_passphrase: None,
    signature_name: None,
    signature_email: None,
};

let repo = Repository::clone_or_open(&configuration).unwrap();

Fields
repository: Repository

The repository struct that is wrapped. Use this to perform operations directly on the repository

Methods

impl<'repo> Repository<'repo>

fn new(
repository: Repository,
configuration: &'repo RepositoryConfiguration
) -> Repository<'repo>

Create a new struct based off a git2::Repository and the corresponding configuration

Examples

extern crate git2;
extern crate fusionner;
extern crate tempdir;

use fusionner::RepositoryConfiguration;
use fusionner::git::Repository;
use tempdir::TempDir;

let td = TempDir::new("checkout").unwrap();
let configuration = RepositoryConfiguration {
    uri: "https://github.com/lawliet89/fusionner.git".to_string(),
    checkout_path: td.path().to_str().unwrap().to_string(),
    fetch_refspecs: vec![],
    push_refspecs: vec![],
    username: None,
    password: None,
    key: None,
    key_passphrase: None,
    signature_name: None,
    signature_email: None,
};

let repo = git2::Repository::clone(&configuration.uri, &configuration.checkout_path)
    .and_then(|repo| Ok(Repository::new(repo, &configuration)))
    .unwrap();

fn clone_or_open(
repo_details: &'repo RepositoryConfiguration
) -> Result<Repository<'repo>, Error>

Convenience method to create a new struct by first attempting to open a repository at the checkout path configured, and failing that will attempt to clone from the URI configured.

fn open(
repo_details: &'repo RepositoryConfiguration
) -> Result<Repository<'repo>, Error>

Convenience method to create a new struct by attempting to open a repository at the checkout path configured

fn clone(
repo_details: &'repo RepositoryConfiguration
) -> Result<Repository<'repo>, Error>

Convenience method to create a new struct by attempting to clone a repository at a uri to the checkout path configured

fn remote(&self, remote: Option<&str>) -> Result<Remote, Error>

Returns a Remote struct for the remote with the given name. Defaults to the origin remote.

fn signature(&self) -> Result<Signature, Error>

Returns a signature struct for use in operations like commit. Will first check if this is available in the configuration. If not, we will attempt to find the global git configured signature. Failing that, we will use some default fusionner signature