Module fusionner::merger

Structs

Merge

Merger

Note
A Note is stored for each commit on the topic branches' current head

Enums

MergeReferenceNamer
The default namer will create a reference at refs/fusionner/{reference}/{target} where {target} is the target reference, and `{reference}~ is the reference that is being merged into target.

ShouldMergeResult
Enum returned by Merger::should_merge depending on the state of affairs

Type Definitions

MergeReferenceNamerCallback
Fn(reference: &str, target_reference: &str, oid: git2::Oid, target_oid: git2::Oid) -> String