Enum fusionner::merger::MergeReferenceNamer

pub enum MergeReferenceNamer<'cb> {
    Default,
    Custom(Box<MergeReferenceNamerCallback<'cb>>),
}

Customise how the merge references are named.

Variants
Default

The default namer will create a reference at refs/fusionner/{reference}/{target} where {target} is the target reference, and `{reference}~ is the reference that is being merged into target.

Note: The default namer will strip everything until the last / so make sure you don't use / in your branch names to avoid collision.

Custom(Box<MergeReferenceNamerCallback<'cb>>)

Use a function that will return the name of the merge reference

Example:

use fusionner::merger::*;
let cb = Box::new(move |reference : &str, target_ref : &str, _oid : _, _target_oid : _| {
                    format!("refs/merge/{}/{}", target_ref, reference)
                  });
let namer = MergeReferenceNamer::Custom(cb);

Methods

impl<'cb> MergeReferenceNamer<'cb>

fn resolve(
&self,
reference: &str,
target_reference: &str,
oid: Oid,
target_oid: Oid
) -> String

Returns the name of the merge reference based on the rules provided.

fn add_default_refspecs(remote: &Remote) -> Result<(), Error>

Add the refspecs used in the Default namer to the remote.