Enum fusionner::merger::ShouldMergeResult

pub enum ShouldMergeResult {
    Merge(Option<Note>),
    ExistingMergeInSameTargetReference(Note),
    ExistingMergeInDifferentTargetReference {
        note: Note,
        merges: Vec<Merge>,
        proposed_merge: Merge,
    },
}

Enum returned by Merger::should_merge depending on the state of affairs

Variants
Merge(Option<Note>)

A merge is required. This variant includes a Note for the commit, if one exists. If one exists, a Merge should be appended to this. Otherwise, a new Note with one Merge should be created

ExistingMergeInSameTargetReference(Note)

An up to date (i.e. target reference OID matches) merge commit is stored in the Note with the target reference exists. Nothing to do.

ExistingMergeInDifferentTargetReference

An up to date (i.e. target reference OID matches) merge commit is stored in the Note. The proposed_merge should be merged into the note

Fields of ExistingMergeInDifferentTargetReference

note: Note

Note found for commit

merges: Vec<Merge>

All the merges for this commit

proposed_merge: Merge

A proposed Merge that should be appended to the Note