Struct fusionner::merger::Merge

pub struct Merge {
    pub merge_oid: String,
    pub target_parent_oid: String,
    pub target_parent_reference: String,
    pub parents_oid: Vec<String>,
    pub merge_reference: String,
}

Denotes a single Merge commit for some target reference. Stored in a Note.

Fields
merge_oid: String

The OID for the merge commit

target_parent_oid: String

The oid on the target branch that was used for the merge commit

target_parent_reference: String

Reference of the target branch parent

parents_oid: Vec<String>

Any other merge parents, other than the target parent

merge_reference: String

The reference for the merge commit

Methods

impl Merge

fn new(
merge_oid: Oid,
target_parent_oid: Oid,
target_parent_reference: &str,
parents: &[Oid],
merge_reference: &str
) -> Merge

Creates a new Merge.