Struct fusionner::merger::Merger

pub struct Merger<'repo, 'cb> { /* fields omitted */ }

Methods

impl<'repo, 'cb> Merger<'repo, 'cb>

fn new(repository: &'repo Repository<'repo>,
remote: Option<&str>,
namespace: Option<&str>,
merge_reference_namer: Option<MergeReferenceNamer<'cb>>)
-> Result<Merger<'repo, 'cb>, Error>

fn add_note_refspecs(&self) -> Result<(), Error>

Add refspecs to a remote to fetch/push commit notes, specific for fusionner

fn fetch_notes(&mut self) -> Result<(), Error>

fn find_note(&self, oid: Oid) -> Result<Note, Error>

Find note for commit. Make sure you have fetched them first

fn add_note(&self, note: &Note, oid: Oid) -> Result<Oid, Error>

Returns OID of note

fn should_merge(&self,
oid: Oid,
target_oid: Oid,
reference: &str,
target_reference: &str)
-> ShouldMergeResult

Determine if a merge should be made

fn merge(&self,
oid: Oid,
target_oid: Oid,
reference: &str,
target_reference: &str)
-> Result<Merge, Error>

Performs a merge and return a Merge entry intended for oid

fn check_and_merge(&mut self,
oid: Oid,
target_oid: Oid,
reference: &str,
target_ref: &str,
push: bool)
-> Result<(Merge, ShouldMergeResult), Error>

Convenience method to check if a merge is required, and merge if needed. Will fetch remote merge references. Will push, if desired

fn notes_reference(&self) -> String