Struct fusionner::RepositoryConfiguration

pub struct RepositoryConfiguration {
    pub uri: String,
    pub checkout_path: String,
    pub fetch_refspecs: Vec<String>,
    pub push_refspecs: Vec<String>,
    pub username: Option<String>,
    pub password: Option<Password>,
    pub key: Option<String>,
    pub key_passphrase: Option<Password>,
    pub signature_name: Option<String>,
    pub signature_email: Option<String>,
}

Configuration struct for the repository

Fields
uri: String

URI to the repository remote.

checkout_path: String

Path to checkout the repository to locally

fetch_refspecs: Vec<String>

Fetch refspecs to add to the repository being checked out. You should make sure that the references you are watching is covered by these refspecs

push_refspecs: Vec<String>

Push refspecs to add to the repository being checked out. This is not really useful at the moment because fusionner will always merge to refs/fusionner/* and automatically adds the right refspecs for you

username: Option<String>

Username to authenticate with the remote

password: Option<Password>

Password to authenticate with the remote

key: Option<String>

Path to private key to authenticate with the remote. If the remote requrests for a key and this is not specified, we will try to request the key from ssh-agent

key_passphrase: Option<Password>

Passphrase to the private key for authentication

signature_name: Option<String>

The name to create merge commits under. If unspecified, will use the global configuration in Git. Otherwise we will use some generic one

signature_email: Option<String>

The email to create merge commits under. If unspecified, will use the global configuration in Git. Otherwise we will use some generic one