Struct fusionner::git::Remote

pub struct Remote<'repo> {
    pub remote: Remote<'repo>,
    // some fields omitted
}

Wraps around a git2::Remote struct and offers convenience methods

Examples

extern crate tempdir;
extern crate fusionner;
use fusionner::RepositoryConfiguration;
use fusionner::git::Repository;
use tempdir::TempDir;

let td = TempDir::new("checkout").unwrap();
let configuration = RepositoryConfiguration {
    uri: "https://github.com/lawliet89/fusionner.git".to_string(),
    checkout_path: td.path().to_str().unwrap().to_string(),
    fetch_refspecs: vec![],
    push_refspecs: vec![],
    username: None,
    password: None,
    key: None,
    key_passphrase: None,
    signature_name: None,
    signature_email: None,
};

let repo = Repository::clone_or_open(&configuration).unwrap();
let remote = repo.remote(None);

Fields
remote: Remote<'repo>

The wrapped remote

Methods

impl<'repo> Remote<'repo>

fn disconnect(&mut self)

Disconnect from the remote

fn name(&self) -> Option<&str>

Returns the name of a remote. Will return None for annonymous remotes

fn refspecs(&self) -> Refspecs

Returns the list of refspecs configured for the remote

fn remote_ls(&mut self) -> Result<Vec<RemoteHead>, Error>

Performs a git ls-remote operation

fn head(&mut self) -> Result<Option<String>, Error>

Get the remote reference of renote HEAD (i.e. default branch)

fn resolve_head(heads: &[RemoteHead]) -> Option<String>

Resolve the remote HEAD (i.e. default branch) from a list of heads and return the remote reference

fn fetch(&mut self, refspecs: &[&str]) -> Result<(), Error>

Fetch the list of refspecs from the remote.

fn push(&mut self, refspecs: &[&str]) -> Result<(), Error>

Attempt to push to the remote for the given list of refspecs

fn generate_refspec(&self, src: &str, force: bool) -> Result<String, String>

For a given local reference, generate a refspec for the remote with the same path on remote i.e. refs/pulls/* --> refs/pulls/*:refs/remotes/origin/pulls/*

fn add_refspec(&self, refspec: &str, direction: Direction) -> Result<(), Error>

Add refspec for the remote, if they don't exist.

fn add_refspecs(
&self,
refspecs: &[&str],
direction: Direction
) -> Result<(), Error>

Convenience method to add multiple refspecs at once

fn resolve_target_ref(
&mut self,
target_ref: Option<&str>
) -> Result<String, Error>

Given a reference string, attempt to find a matching remote reference. If None is provided, will attempt to resolve the remote's HEAD (usually refs/heads/master)

fn find_matching_refspec<'a>(
refspecs: Refspecs<'a>,
direction: Direction,
refspec: &str
) -> Option<Refspec<'a>>

Find if a refspec exists in a list of refspecs, usually retrieved from a repository or a remote

fn direction_eq(left: &Direction, right: &Direction) -> bool

Convenience method to check if two git2::Direction enums are equal