Module fusionner::merger

Contains the logic to create merge commits.

Metadata from fusionner are stored as git object notes. Read this for an introduction. Each commit handled by fusionner will have a Note serialised to TOML stored under some configured namespace. The default is fusionner. Each note may contain multiple Merges.

You can list the notes in your command line by running git notes --ref fusionner list.

Terminology

oid and reference refer to the OID and Git reference for the commit that you are intending to merge into some target_oid and target_reference. The former pair usually corresponds to some topic branch while the latter pair is usually some default branch (i.e. master).

Structs

Merge
Denotes a single Merge commit for some target reference. Stored in a Note.

Merger
Contains the logic to create, and manage merge commits

Note
A Note is stored for each commit on the topic branches' current head

Enums

MergeReferenceNamer
Customise how the merge references are named.

ShouldMergeResult
Enum returned by Merger::should_merge depending on the state of affairs

Type Definitions

MergeReferenceNamerCallback
Type for callback implementing custom merge reference naming. The meanings for the input arguments are: Fn(reference: &str, target_reference: &str, oid: git2::Oid, target_oid: git2::Oid) -> String

Merges
A HashMap of Merge where the key is a String corresponding to the target_reference of the merge. This ensures that only one merge commit per target_reference is tracked.