Struct fusionner::merger::Merger

pub struct Merger<'repo, 'cb> { /* fields omitted */ }

Contains the logic to create, and manage merge commits

Methods

impl<'repo, 'cb> Merger<'repo, 'cb>

fn new(
repository: &'repo Repository<'repo>,
remote_name: Option<&str>,
namespace: Option<&str>,
merge_reference_namer: Option<MergeReferenceNamer<'cb>>
) -> Result<Merger<'repo, 'cb>, Error>

Create a new merger.

Provide the name of the remote to use, or the default (usually origin) will be used.

Examples

use fusionner::RepositoryConfiguration;
use fusionner::git::Repository;
use fusionner::merger::*;

let configuration = RepositoryConfiguration {
    uri: "https://github.com/lawliet89/fusionner.git".to_string(),
    checkout_path: "/tmp/checkout".to_string(),
    fetch_refspecs: vec![],
    push_refspecs: vec![],
    username: None,
    password: None,
    key: None,
    key_passphrase: None,
    signature_name: None,
    signature_email: None,
};

let repo = Repository::clone_or_open(&configuration).unwrap();
let cb = Box::new(move |reference : &str, target_ref : &str, _oid : _, _target_oid : _| {
                    format!("refs/merge/{}/{}", target_ref, reference)
                  });
let namer = MergeReferenceNamer::Custom(cb);
let merger = Merger::new(&repo, None, None, Some(namer));

fn add_note_refspecs(&self) -> Result<(), Error>

Add refspecs to a remote to fetch/push commit notes, specific for fusionner This is based on the namespace provided when creating a new Merger.

fn fetch_notes(&mut self) -> Result<(), Error>

Fetch notes based on the namespace provided when creating a new Merger from the remote configured

fn find_note(&self, oid: Oid) -> Result<Note, Error>

Find note for commit. Make sure you have fetched them first

fn add_note(&self, note: &Note, oid: Oid) -> Result<Oid, Error>

Add note for the OID. Will serialise to TOML before storage. Returns OID of note.

fn should_merge(
&self,
oid: Oid,
target_oid: Oid,
reference: &str,
target_reference: &str
) -> ShouldMergeResult

Determine if a merge should be made, based on the rules below.

In general, you should prefer to use the convenience function check_and_merge instead which will do everything for you. For usage of this function, refer to the source code of check_and_merge.

Find the note for the oid in question.

If no note could be found, return ShouldMergeResult::Merge(None).

If a note could be found, find from its list of Merges if any merge with the target_oid could be found.

If none could be found, return ShouldMergeResult::Merge(Some(note)) where note is the Note found.

If the found Merge has its target_reference match, then we will return ShouldMergeResult::ExistingMergeInSameTargetReference.

Otherwise, we will construct a proposed_merge and return ShouldMergeResult::ExistingMergeInDifferentTargetReference.

fn merge(
&self,
oid: Oid,
target_oid: Oid,
reference: &str,
target_reference: &str
) -> Result<Merge, Error>

Performs a merge and return a Merge entry intended for oid. You should then add the Merge into the Note for oid.

In general, you should prefer to use the convenience function check_and_merge instead which will do everything for you. For usage of this function, refer to the source code of check_and_merge.

fn check_and_merge(
&mut self,
oid: Oid,
target_oid: Oid,
reference: &str,
target_ref: &str,
push: bool
) -> Result<(Merge, ShouldMergeResult), Error>

Convenience function to check if a merge is required, and merge if needed. Will fetch remote merge references. Will push, if desired. This function calls both should_merge and merge.

fn notes_reference(&self) -> String

Returns the reference for the notes that fusionner will create, based on the namespace provided when creating a new Merger.